flo-opt: An optimizer for Flo graphs
This tool performs optimizations on Flo files and produces Flo files.
The general idea here is that some sorts of optimizations can be done
without the need to actually generate code for a particular target but
can instead simply be done at the granularity of individual Flo
operations.

The purpose of this tool is to construct a database of these
optimizations that can be used to target different Flo emulation
machines.